Previous revision to native-thunk-generation wasn't 64-bit safe. AbcThunkGen (in asc.jar) was revised to output thunks that should now work for 64-bit builds ( note, you need rev 2826 or later of asc.jar, please visit http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc for source ).
While I was at it I resurrected the 64-bit XCode build for Mac; we now (mostly)
run properly in non-MIR mode on 64-bit Mac. As a result, fixed a few minor
32/64 related warnings.

r=wsharp

+++ core/Sampler.cpp
@@ -118,7 +118,7 @@ namespace avmplus
 			write(p, depth);
 			while(csn)
 			{
-				write(p, (uint32)csn->info);
+				write(p, (uintptr_t)csn->info);
 				// FIXME: can filename can be stored in the AbstractInfo?
 				write(p, csn->filename);
 				write(p, csn->linenum);
@@ -179,7 +179,7 @@ namespace avmplus
 
 		if(typeOrVTable < 7 && core->codeContext() && core->codeContext()->domainEnv()) {
 			// and in toplevel
-			typeOrVTable |= (uint32)core->codeContext()->domainEnv()->toplevel();
+			typeOrVTable |= (uintptr_t)core->codeContext()->domainEnv()->toplevel();
 		}
 
 		writeRawSample(NEW_OBJECT_SAMPLE);
